+++ pkg/errno/code.go
@@ -76,6 +76,6 @@ const (
 	InternalETCDErrorCode      = 50019 // ETCD错误
 	InternalTraceErrorCode     = 50020 // Trace错误
 
-	// paper的SuccessCode是2000, 不是 10000
+	// paper在旧版Android中的SuccessCode是2000，用作兼容
 	SuccessCodePaper = 2000
 )
